[WIP] need to use .env for live deployment #82
cause docker-compose up is not willing to use .env.docker

it would complain as follows:

$ docker-compose -f docker-compose.yml -f docker-compose.live.yml config
WARNING: The APACHE_LOG_DIR variable is not set. Defaulting to a blank string.
WARNING: The PHP_LOG_DIR variable is not set. Defaulting to a blank string.
WARNING: The WAYPOINTS_IMAGES_DIR variable is not set. Defaulting to a blank string.
WARNING: The OUTER_WEB_PORT variable is not set. Defaulting to a blank string.
ERROR: The Compose file './docker-compose.live.yml' is invalid because:
services.web.ports is invalid: Port ranges don't match in length